Removed Area FK in Advertisement, add new column area, country, city

+++ EzRental/Controllers/AdvertisementController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using EzRental.Data;
 using EzRental.Models;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace EzRental.Controllers
 {
@@ -30,7 +31,7 @@ namespace EzRental.Controllers
                 return Problem("Server ran into an unexpected error");
             }
 
-            var advertisements = await _context.Advertisement.Include(ad => ad.Area).ToListAsync();
+            var advertisements = await _context.Advertisement.ToListAsync();
                 //Include(ad => ad.Area.city).Include(ad => ad.Area.city.Country).ToListAsync();
 
             if (advertisements.Count > 0) { return Ok(advertisements); } else return NotFound(); 
@@ -40,57 +41,82 @@ namespace EzRental.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<Advertisement>> GetAdvertisement(int id)
         {
-          if (_context.Advertisement == null)
-          {
-              return NotFound();
-          }
-
-            var advertisement = await _context.Advertisement.Include(ad => ad.Rent).Include(ad => ad.Rent.Room).
-                Include(ad => ad.Rent.Renter).Include(ad => ad.Area).
-                  Include(ad => ad.Area.city).Include(ad => ad.Area.city.Country).FirstAsync(ad => ad.AdId == id);
-
-            if(advertisement == null)
+            try
             {
-                return NotFound();
-            }
-
-            AdvertisementWrapper advertisementWrapper = new AdvertisementWrapper();
-            advertisementWrapper.advertisement = advertisement;
-            advertisementWrapper.facilties = new List<Facilties>();
-
-            var facilities = _context.AdFacility.Include(af => af.Facility).Where(af => af.AdId == id).ToList();
-
-            if(facilities != null)
-                foreach(var facility in facilities)
-                {   
-                    if(facility.Facility != null)
-                        advertisementWrapper.facilties.Add(facility.Facility);
+                if (_context.Advertisement == null)
+                {
+                    return NotFound();
                 }
 
-            return Ok(advertisementWrapper);
+                var advertisement = await _context.Advertisement.Include(ad => ad.Rent).Include(ad => ad.Rent.Room).
+                    Include(ad => ad.Rent.Renter).FirstAsync(ad => ad.AdId == id);
+
+                if (advertisement == null)
+                {
+                    return NotFound();
+                }
+
+                AdvertisementWrapper advertisementWrapper = new AdvertisementWrapper();
+                advertisementWrapper.advertisement = advertisement;
+                advertisementWrapper.facilties = new List<Facilties>();
+
+                var facilities = _context.AdFacility.Include(af => af.Facility).Where(af => af.AdId == id).ToList();
+
+                if (facilities != null)
+                    foreach (var facility in facilities)
+                    {
+                        if (facility.Facility != null)
+                            advertisementWrapper.facilties.Add(facility.Facility);
+                    }
+
+                return Ok(advertisementWrapper);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         // PUT: api/Advertisement/5
         [HttpPut("{id}")]
         public async Task<IActionResult> PutAdvertisement(int id, AdvertisementWrapper advertisementWrapper)
         {
+            if (_context == null)
+                return BadRequest("Server ran into an error");
+
+            if (advertisementWrapper.advertisement == null || advertisementWrapper.facilties == null)
+                return BadRequest("Server ran into an error");
+
+            if (!AdvertisementExists(id))
+                return NotFound("Advertisement not Found");
+
             try
             {
                 Advertisement advertisement = advertisementWrapper.advertisement;
                 List<Facilties> facilities = advertisementWrapper.facilties;
 
+                if (advertisement.Rent == null || advertisement.Rent.Room == null)
+                    return BadRequest("Server ran into an error");
 
-                if (id != advertisement.AdId)
-                {
-                    return BadRequest();
-                }
+                //set advertisement Id
+                advertisement.AdId = id;
 
+                //set room id to be tracked
+                int rent_id = await _context.Advertisement.Where(ad => ad.AdId == id).
+                    Select(ad => ad.RentId).FirstAsync();
+                int room_id = await _context.Rent.Where(r => r.RentId == rent_id).
+                    Select(r => r.RoomId).FirstAsync();
+
+                advertisement.Rent.Room.RoomId = room_id;
+                advertisement.RentId = rent_id;
+
+                
                 _context.Entry(advertisement.Rent.Room).State = EntityState.Modified;
                 _context.Entry(advertisement).State = EntityState.Modified;
 
                 try
                 {
-                    await _context.SaveChangesAsync();
+                    _context.SaveChanges();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -104,44 +130,45 @@ namespace EzRental.Controllers
                     }
                 }
 
-                var existing_facilities = _context.AdFacility.Where(ad => ad.AdId == id).Select(ad => ad.FacilityId).ToListAsync();
+                var existing_facilities = _context.AdFacility.Where(ad => ad.AdId == id).
+                    Select(ad => ad.FacilityId).ToListAsync();
 
-                List<int> facility_ids = new List<int>();
-                foreach(var facility in facilities)
-                {
-                    facility_ids.Add(facility.FacilityId);
-                }
+
+                var facility_ids = facilities.Select(obj => obj.FacilityId).ToList();
                 
-                // add new facilities
-                foreach(var facility in facilities) 
+                // add new facilities is the previous state of object does not already contains it
+                foreach(var facility in facility_ids) 
                 {
-                    if (!existing_facilities.Result.Contains(facility.FacilityId))
+                    if (!existing_facilities.Result.Contains(facility))
                     {
                         Console.WriteLine("New Facility");
-                        _context.AdFacility.Add(new AdFacility { FacilityId = facility.FacilityId, AdId = id });
+                        _context.AdFacility.Add(new AdFacility { FacilityId = facility, AdId = id });
                     }
                     else
                     {
-                        Console.WriteLine($"Same facility {facility.FacilityId} detected");
+                        Console.WriteLine($"Same facility {facility} detected");
                     }
                 }
 
-                // remove facilities
+                // remove facilities is the current facilties object does not contains existing facilities
                 foreach (var facility in existing_facilities.Result)
                 {
                     if (!facility_ids.Contains(facility))
                     {
                         Console.WriteLine($"Remove Facility {facility}");
                         AdFacility adfacility = await _context.AdFacility.FirstOrDefaultAsync(af => af.FacilityId == facility && af.AdId == id);
+                        
+                        if (adfacility == null)
+                            continue;
+
                         _context.AdFacility.Remove(adfacility);
                     }
                 }
 
-
                 await _context.SaveChangesAsync();
 
 
-                return NoContent();
+                return Ok();
             }
             catch(Exception e) 
             {
@@ -167,27 +194,18 @@ namespace EzRental.Controllers
                 Advertisement advertisement = advertisementWrapper.advertisement;
                 List<Facilties> facilities = advertisementWrapper.facilties;
 
-                _context.Room.Add(advertisement.Rent.Room);
-                _context.SaveChanges();
-
-                advertisement.Rent.RoomId = advertisement.Rent.Room.RoomId;
                 advertisement.Rent.Renter = null;
-                _context.Rent.Add(advertisement.Rent);
-                _context.SaveChanges();
-                
-                advertisement.RentId = advertisement.Rent.RentId;
-                advertisement.Rent = null;
-                advertisement.Area = null;
                 _context.Advertisement.Add(advertisement);
-                _context.SaveChanges();
-
+                await _context.SaveChangesAsync();
+              
                 foreach(var facility in facilities)
                 {
-                    _context.AdFacility.Add(new AdFacility { FacilityId = facility.FacilityId,AdId=advertisement.AdId });
+                    _context.AdFacility.Add(new AdFacility { FacilityId = facility.FacilityId, AdId=advertisement.AdId });
                 }
 
                 await _context.SaveChangesAsync();
-                return CreatedAtAction("GetAdvertisement", new { id = advertisement.AdId });
+                return Created("Advertisement Added", new { id = advertisement.AdId});
+                
             }
             catch (Exception e)
             {
@@ -206,7 +224,7 @@ namespace EzRental.Controllers
             }
 
             // get advertisement data
-            var advertisement = await _context.Advertisement.Include(a => a.Rent).Include(r => r.Rent.Room).FirstAsync(a => a.AdId == id);
+            var advertisement = await _context.Advertisement.Include(a => a.Rent).Include(a => a.Rent.Room).FirstAsync(a => a.AdId == id);
             
             if (advertisement == null)
             {
